Fix async chunk send  (#252)
If a chunk is already in the sentChunks set, we simply return early instead of throwing an exception.

Made moonrise$addReceivedChunk call to gracefully handle the case where a chunk is already marked as received.

If the chunk was already received, we remove the chunk from our sent list and return early.