Remove toggleable async catcher
This option is useless, especially provided as an config option,
since the normal server owners have no reasons to use it.
And even for developers, this option is still unnecessary to use.
It's also a very simple toggle which can be done in
local stage during debugging or fixing specific issue.